* Some better type defs. Also consider reworking harvesting as it currently throtles the game...

+++ src/game/algorithms/PYAPI/PyAPI.h
@@ -40,7 +40,7 @@ public:
 	
 	static void start();
 
-	PyAPI(uint16_t gameID, uint16_t playerID);
+	PyAPI(uint8_t gameID, uint8_t playerID);
 	std::string scriptName;
